Feat: toast 튀어오르게 수정
