--- src/components/common/Button/LinkBox.module.css
@@ -1,7 +1,0 @@
-.linkContainer {
-  display: flex;
-  align-items: center;
-  margin: 0;
-  padding: 0 1.2rem 0 0;
-  list-style: none;
-}
