+++ src/components/common/Button/ShareBox.module.css
@@ -1,4 +1,4 @@
 .shareBox {
   display: flex;
-  padding: 12px 16px 12px 16px;
+  column-gap: 1.2rem;
 }
