+++ src/components/common/Button/Toast.jsx
@@ -1,9 +1,15 @@
+import { useEffect } from 'react';
 import styles from './Toast.module.css';
 
-export default function Toast({ pop }) {
-  return (
-    <div className={styles.toast}>
-      <p className={styles.message}>{pop}</p>
-    </div>
-  );
+export default function Toast({ setToast, text, toastTime }) {
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setToast(false);
+    }, toastTime);
+    return () => {
+      clearTimeout(timer);
+    };
+  }, [setToast]);
+
+  return <p className={styles.toast}>{text}</p>;
 }
